+++ modules/apps/login/login-authentication-openid-web/src/main/java/com/liferay/login/authentication/openid/web/internal/portlet/action/OpenIdLoginMVCActionCommand.java
@@ -29,9 +29,9 @@ import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.security.sso.openid.OpenIdServiceException;
 import com.liferay.portal.security.sso.openid.OpenIdServiceHandler;
-import com.liferay.portal.security.sso.openid.StrangersNotAllowedException;
+import com.liferay.portal.security.sso.openid.exception.OpenIdServiceException;
+import com.liferay.portal.security.sso.openid.exception.StrangersNotAllowedException;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
