+++ modules/apps/portal-security-sso/portal-security-sso-openid-api/bnd.bnd
@@ -4,4 +4,5 @@ Bundle-Version: 3.0.0
 Export-Package:\
 	com.liferay.portal.security.sso.openid,\
 	com.liferay.portal.security.sso.openid.configuration,\
-	com.liferay.portal.security.sso.openid.constants
+	com.liferay.portal.security.sso.openid.constants,\
+	com.liferay.portal.security.sso.openid.exception
