+++ modules/apps/portal-security-sso/portal-security-sso-openid-api/src/main/java/com/liferay/portal/security/sso/openid/exception/OpenIdServiceException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.openid;
+package com.liferay.portal.security.sso.openid.exception;
 
 import com.liferay.portal.kernel.exception.PortalException;
