+++ modules/apps/portal-security-sso/portal-security-sso-openid-api/src/main/resources/com/liferay/portal/security/sso/openid/exception/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
