+++ modules/apps/portal-security-sso/portal-security-sso-openid-api/src/main/resources/com/liferay/portal/security/sso/openid/packageinfo
@@ -1,1 +1,1 @@
-version 2.1.0
+version 3.0.0
